refactor: co-locate route components
