+++ src/app/account/_components/AgentManagement/index.ts
@@ -1,0 +1,2 @@
+export { default as AgentPromptForm } from './AgentPromptForm'
+export { default as AgentWalletInfo } from './AgentWalletInfo'
