+++ src/app/daos/_components/DaoActions/DaoBuyToken.tsx
@@ -3,7 +3,7 @@
 import type React from "react";
 import { useState } from "react";
 import { Button } from "@/components/ui/button";
-import { TokenPurchaseModal } from "@/components/daos/TokenPurchaseModal";
+import { TokenPurchaseModal } from "./TokenPurchaseModal";
 
 interface DAOBuyTokenProps {
   daoId: string;
