+++ src/app/daos/_components/DaoActions/TokenPurchaseModal.tsx
@@ -18,7 +18,7 @@ import {
   RotateCcw,
 } from "lucide-react";
 import { Loader } from "@/components/reusables/Loader";
-import { TokenBuyInput } from "@/components/daos/DaoBuy";
+import { TokenBuyInput } from "./DaoBuy";
 import type { ApiResponse } from "@/services/tool.service";
 import { useAuth } from "@/hooks/useAuth";
 import { useWalletStore } from "@/store/wallet";
