+++ src/app/daos/_components/DaoList/index.ts
@@ -1,0 +1,3 @@
+export { default as AllDaos } from './AllDaos'
+export { default as DaosTable } from './DaosTable'
+export { DAOCard } from './DaoCard'
