+++ src/app/daos/_components/index.ts
@@ -1,0 +1,3 @@
+export * from './DaoList'
+export * from './DaoActions'
+export * from './shared'
