+++ src/app/proposals/[id]/_components/ProposalDetails/ProposalDetails.tsx
