+++ src/app/proposals/[id]/_components/ProposalDetails/index.ts
@@ -1,0 +1,1 @@
+export { default as ProposalDetails } from './ProposalDetails'
