+++ src/app/proposals/[id]/page.tsx
@@ -5,11 +5,11 @@ import { useQuery } from "@tanstack/react-query";
 import { Suspense } from "react";
 import { ProposalWithDAO } from "@/types";
 import { fetchProposalById } from "@/services/dao.service";
-import ProposalDetails from "@/components/proposals/ProposalDetails";
+import { ProposalDetails } from "./_components/ProposalDetails";
 import { Button } from "@/components/ui/button";
 import { ArrowLeft, ExternalLink } from "lucide-react";
 import { Badge } from "@/components/ui/badge";
-import { useVotingStatus } from "@/components/proposals/TimeStatus";
+import { useVotingStatus } from "../_components";
 import { safeNumberFromBigInt, safeString } from "@/utils/proposal";
 import { format } from "date-fns";
 import Link from "next/link";
