+++ src/app/proposals/_components/ProposalActions/index.ts
@@ -1,0 +1,3 @@
+export { default as DAOProposal } from './DAOProposal'
+export { default as DAOSendProposal } from './DAOSendProposal'
+export { default as DAOVetoProposal } from './DAOVetoProposal'
