+++ src/app/proposals/_components/ProposalList/index.ts
@@ -1,0 +1,4 @@
+export { default as AllProposals } from './AllProposals'
+export { default as ProposalCard } from './ProposalCard'
+export { default as ProposalsList } from './ProposalsList'
+export { default as ProposalList } from './ProposalList'
