+++ src/app/proposals/_components/ProposalVoting/index.ts
@@ -1,0 +1,3 @@
+export { default as VoteStatusChart } from './VoteStatusChart'
+export { default as VotingProgressChart } from './VotingProgressChart'
+export { default as VotesTable } from './VotesTable'
