+++ src/app/proposals/_components/index.ts
@@ -1,0 +1,4 @@
+export * from './ProposalList'
+export * from './ProposalActions'
+export * from './ProposalVoting'
+export * from './shared'
