+++ src/app/proposals/_components/shared/BlockCounter.tsx
