+++ src/app/proposals/_components/shared/BlockVisual.tsx
