+++ src/app/proposals/_components/shared/CopyButton.tsx
