+++ src/app/proposals/_components/shared/LabeledField.tsx
