+++ src/app/proposals/_components/shared/MessageDisplay.tsx
