+++ src/app/proposals/_components/shared/ProposalMetrics.tsx
