+++ src/app/proposals/_components/shared/TimeStatus.tsx
