+++ src/layouts/index.ts
@@ -1,3 +1,3 @@
 export { DAOLayout } from "./DAOLayout";
-export { DAOHeader } from "@/components/daos/DAOHeader";
-export { MissionContent } from "@/components/daos/MissionContent";
+export { DAOHeader } from "@/app/daos/_components";
+export { MissionContent } from "@/app/daos/_components";
